fix(avatars): refactor types and interfaces
