+++ packages/avatars/src/elements/Avatar.tsx
@@ -5,27 +5,11 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-import React, { Children, forwardRef, HTMLAttributes } from 'react';
+import React, { Children, forwardRef } from 'react';
 import PropTypes from 'prop-types';
+import { IAvatarProps, SIZE, STATUS } from '../types';
 import { StyledAvatar } from '../styled';
 import { Text } from './components/Text';
-
-export interface IAvatarProps extends HTMLAttributes<HTMLElement> {
-  /** Sets the avatar background color */
-  backgroundColor?: string;
-  /** Sets the color for child SVG or `Avatar.Text` components */
-  foregroundColor?: string;
-  /** Provides surface color for an avatar placed on a non-white background */
-  surfaceColor?: string;
-  /** Applies system styling for representing objects, brands, or products */
-  isSystem?: boolean;
-  /** Specifies the avatar size */
-  size?: 'extraextrasmall' | 'extrasmall' | 'small' | 'medium' | 'large';
-  /** Applies status styling */
-  status?: 'available' | 'away';
-  /** Sets the badge text and applies active styling */
-  badge?: string | number;
-}
 
 const AvatarComponent = forwardRef<HTMLElement, IAvatarProps>(
   (
@@ -72,8 +56,8 @@ AvatarComponent.propTypes = {
   surfaceColor: PropTypes.string,
   isSystem: PropTypes.bool,
   badge: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  size: PropTypes.oneOf(['extraextrasmall', 'extrasmall', 'small', 'medium', 'large']),
-  status: PropTypes.oneOf(['available', 'away'])
+  size: PropTypes.oneOf(SIZE),
+  status: PropTypes.oneOf(STATUS)
 };
 
 AvatarComponent.defaultProps = {
